fix compilation with simde
for arm we can omit all the magic that compiles for different kinds of
simd instruction sets like avx, avx2 is done for intel. For arm we use
simde with the avx2 instruction set versions.

Hence, this commit adds two switches to the cmake interface:
* `USE_SIMDE_ON_ARM`: to activate using SIMDE to translate to whatever
  target architecture (that is supported)
* `APPLE_ARM64`: for explicit clang targeting m1/m2 compilation flags
  foreseeing that it might be slightly different for arch64

Then, throughout the code, there are now splattered `_USE_SIMDE_ON_ARM_`
conditionals. E.g., the various `*_init` functions that would check for
certain processor capabilities at runtime to activate the most effective
code path. These all point to the avx2 versions when SIMDE is active.

All in all not very pretty - might want to clean this up a bit in the
future...