pin clang to 16 on osx
clang 17 seems to have problems with the build, specifically std flags
are not properly recognized via cmake.
Note, I also tried compiling without the explicit setting of the std for
pybind11 and it worked (pybind11 cmake example also doesn't do it
explicitly - so we might also be able to remove this check).